# go/model_delete_task_result.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.11
 * Contact: apiteam@swagger.io
 */

package openapi




type DeleteTaskResult struct {

	Code int64 `json:"code,omitempty"`

	Message string `json:"message,omitempty"`
}

// AssertDeleteTaskResultRequired checks if the required fields are not zero-ed
func AssertDeleteTaskResultRequired(obj DeleteTaskResult) error {
	return nil
}

// AssertDeleteTaskResultConstraints checks if the values respects the defined constraints
func AssertDeleteTaskResultConstraints(obj DeleteTaskResult) error {
	return nil
}