# go/model_get_task_list_result.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.11
 * Contact: apiteam@swagger.io
 */

package openapi




type GetTaskListResult struct {

	Code int64 `json:"code,omitempty"`

	Message string `json:"message,omitempty"`

	Tasks []Task `json:"tasks,omitempty"`
}

// AssertGetTaskListResultRequired checks if the required fields are not zero-ed
func AssertGetTaskListResultRequired(obj GetTaskListResult) error {
	for _, el := range obj.Tasks {
		if err := AssertTaskRequired(el); err != nil {
			return err
		}
	}
	return nil
}

// AssertGetTaskListResultConstraints checks if the values respects the defined constraints
func AssertGetTaskListResultConstraints(obj GetTaskListResult) error {
	for _, el := range obj.Tasks {
		if err := AssertTaskConstraints(el); err != nil {
			return err
		}
	}
	return nil
}