# go/model_get_task_result.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.11
 * Contact: apiteam@swagger.io
 */

package openapi




type GetTaskResult struct {

	Code int64 `json:"code,omitempty"`

	Message string `json:"message,omitempty"`

	Task Task `json:"task,omitempty"`
}

// AssertGetTaskResultRequired checks if the required fields are not zero-ed
func AssertGetTaskResultRequired(obj GetTaskResult) error {
	if err := AssertTaskRequired(obj.Task); err != nil {
		return err
	}
	return nil
}

// AssertGetTaskResultConstraints checks if the values respects the defined constraints
func AssertGetTaskResultConstraints(obj GetTaskResult) error {
	if err := AssertTaskConstraints(obj.Task); err != nil {
		return err
	}
	return nil
}