# go/model_task.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.11
 * Contact: apiteam@swagger.io
 */

package openapi




type Task struct {

	TaskId string `json:"taskId,omitempty"`

	TaskName string `json:"taskName,omitempty"`

	TaskDesc string `json:"taskDesc,omitempty"`

	CreateTime string `json:"createTime,omitempty"`

	TaskType int32 `json:"taskType,omitempty"`

	NotifyType int32 `json:"notifyType,omitempty"`

	NotifyTime string `json:"notifyTime,omitempty"`

	NotifyBeforeTime int32 `json:"notifyBeforeTime,omitempty"`
}

// AssertTaskRequired checks if the required fields are not zero-ed
func AssertTaskRequired(obj Task) error {
	return nil
}

// AssertTaskConstraints checks if the values respects the defined constraints
func AssertTaskConstraints(obj Task) error {
	return nil
}